Refactor summary and json commands
